LibAudio: Move MDCT.h from LibDSP to LibAudio
